Directory as a Python Package
The __init__.py file is a special file used in Python and Django projects to mark a directory as a Python package. In the Class Connect e-learning platform, this file is present in the project folder (e.g., classconnect/) and in each app directory (e.g., student/, courses/, account/, etc.).